// client/gen/event_map_api/v1/event_map_api_connectweb.ts
// @generated by protoc-gen-connect-web v0.6.0 with parameter "target=ts"
// @generated from file event_map_api/v1/event_map_api.proto (package proto.event_map_api.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { CreateEventMapRequest, CreateEventMapResponse, DeleteEventMapRequest, DeleteEventMapResponse, GetAllEventMapsRequest, GetAllEventMapsResponse, GetEventMapRequest, GetEventMapResponse, UpdateEventMapRequest, UpdateEventMapResponse } from "./event_map_api_pb.js";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service proto.event_map_api.v1.EventMapService
 */
export const EventMapService = {
  typeName: "proto.event_map_api.v1.EventMapService",
  methods: {
    /**
     * @generated from rpc proto.event_map_api.v1.EventMapService.CreateEventMap
     */
    createEventMap: {
      name: "CreateEventMap",
      I: CreateEventMapRequest,
      O: CreateEventMapResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc proto.event_map_api.v1.EventMapService.GetEventMap
     */
    getEventMap: {
      name: "GetEventMap",
      I: GetEventMapRequest,
      O: GetEventMapResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc proto.event_map_api.v1.EventMapService.GetAllEventMaps
     */
    getAllEventMaps: {
      name: "GetAllEventMaps",
      I: GetAllEventMapsRequest,
      O: GetAllEventMapsResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc proto.event_map_api.v1.EventMapService.UpdateEventMap
     */
    updateEventMap: {
      name: "UpdateEventMap",
      I: UpdateEventMapRequest,
      O: UpdateEventMapResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc proto.event_map_api.v1.EventMapService.DeleteEventMap
     */
    deleteEventMap: {
      name: "DeleteEventMap",
      I: DeleteEventMapRequest,
      O: DeleteEventMapResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;